+++ host/Source/Control/ActionHandler.h
@@ -86,6 +86,7 @@ private:
     bool preMuteOutputMuted_ = false;
     bool preMuteVstEnabled_ = false;
     bool preMuteRecordingActive_ = false;
+    bool panicRestorePending_ = false;
 
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(ActionHandler)
 };